task creates

+++ app/domains/user/soc_net/entity.go
@@ -4,7 +4,7 @@ import "time"
 
 type UserSocNet struct {
 	ID           int       `json:"id" table_name:"users_soc_nets"`
-	UserId       int64     `json:"user_id"`
+	UserId       int       `json:"user_id"`
 	UserSocNetID string    `json:"soc_net_id"`
 	Created      time.Time `json:"created"`
 	Updated      time.Time `json:"updated"`
